+++ tests/Test/Exception/CustomDateIsNotEqualTo.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\DateTimePrecision\Test\Exception;
 
-final class CustomMomentIsNotBeforeInTimeZone extends \InvalidArgumentException
+final class CustomDateIsNotEqualTo extends \InvalidArgumentException
 {
 }