+++ tests/Test/Exception/CustomTimeIsAfter.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\DateTimePrecision\Test\Exception;
 
-final class CustomMomentIsAfterOrEqualToInTimeZone extends \InvalidArgumentException
+final class CustomTimeIsAfter extends \InvalidArgumentException
 {
 }