+++ tests/Test/Exception/CustomTimeIsAfterOrEqualTo.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\DateTimePrecision\Test\Exception;
 
-final class CustomDateIsEqual extends \InvalidArgumentException
+final class CustomTimeIsAfterOrEqualTo extends \InvalidArgumentException
 {
 }