+++ tests/Test/Exception/CustomTimeIsBeforeOrEqualTo.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\DateTimePrecision\Test\Exception;
 
-final class CustomDateIsNotEqual extends \InvalidArgumentException
+final class CustomTimeIsBeforeOrEqualTo extends \InvalidArgumentException
 {
 }