+++ tests/Test/Exception/CustomTimeIsEqualTo.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\DateTimePrecision\Test\Exception;
 
-final class CustomMomentIsNotEqualInTimeZone extends \InvalidArgumentException
+final class CustomTimeIsEqualTo extends \InvalidArgumentException
 {
 }