+++ tests/Test/Exception/CustomTimeIsNotAfterOrEqualTo.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\DateTimePrecision\Test\Exception;
 
-final class CustomMomentIsBeforeOrEqualToInTimeZone extends \InvalidArgumentException
+final class CustomTimeIsNotAfterOrEqualTo extends \InvalidArgumentException
 {
 }