+++ tests/Test/Exception/CustomTimeIsNotBefore.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\DateTimePrecision\Test\Exception;
 
-final class CustomMomentIsNotBeforeOrEqualToInTimeZone extends \InvalidArgumentException
+final class CustomTimeIsNotBefore extends \InvalidArgumentException
 {
 }