+++ tests/Test/Exception/CustomTimeIsNotBeforeOrEqualTo.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\DateTimePrecision\Test\Exception;
 
-final class CustomMomentIsNotAfterInTimeZone extends \InvalidArgumentException
+final class CustomTimeIsNotBeforeOrEqualTo extends \InvalidArgumentException
 {
 }